# volley-api/volley.openapi.yaml
openapi: 3.0.3
info:
  title: Volley API
  description: API for organizing and joining pickup sports games
  version: 1.0.0
  contact:
    name: Volley Support
    email: support@volley.app
servers:
  - url: https://api.volley.app/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /games:
    get:
      tags:
        - games
      summary: List games
      operationId: listGames
      parameters:
        - name: categories
          in: query
          required: true
          description: Filter by sport categories (can specify multiple)
          schema:
            type: array
            items:
              $ref: '#/components/schemas/GameCategory'
          example: ["soccer"]
        - name: latitude
          in: query
          required: true
          description: Latitude coordinate for location-based search
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 29.7736199
        - name: longitude
          in: query
          required: true
          description: Longitude coordinate for location-based search
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -95.4201315
        - name: radius
          in: query
          description: Search radius in meters (default 16093.4 meters = 10 miles)
          schema:
            type: number
            format: double
            minimum: 0
            default: 16093.4
          example: 16093.4
        - name: timeFilter
          in: query
          description: Filter by time (upcoming, past, or all)
          schema:
            type: string
            enum: [upcoming, past, all]
            default: upcoming
          example: upcoming
        - name: status
          in: query
          description: Filter by game status
          schema:
            $ref: '#/components/schemas/GameStatus'
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of games
          content:
            application/json:
              schema:
                type: object
                required:
                  - games
                properties:
                  games:
                    type: array
                    items:
                      $ref: '#/components/schemas/GameSummary'

    post:
      tags:
        - games
      summary: Create a new game
      operationId: createGame
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
            example:
              category: "soccer"
              title: "Sunday Morning Soccer"
              description: "Casual pickup soccer at O Athletik. All skill levels welcome! Bring cleats and shin guards."
              location:
                name: "O Athletik"
                address: "767 N Shepherd Dr, Houston, TX 77007"
                latitude: 29.7852774
                longitude: -95.4102687
                notes: "Park in the back. Meet at Field 1."
              startTime: "2025-11-28T20:00:00Z"
              durationMinutes: 90
              maxParticipants: 12
              pricing:
                type: "free"
                amountCents: 0
                currency: "USD"
              signupDeadline: "2025-11-27T20:00:00Z"
              skillLevel: "all"
              notes: "We'll split into teams when we get there. Bring water!"
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/{gameId}:
    get:
      tags:
        - games
      summary: Get game details
      operationId: getGame
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - games
      summary: Update game details
      operationId: updateGame
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameRequest'
      responses:
        '200':
          description: Game updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the game owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - games
      summary: Delete a game (hard delete)
      description: Permanently deletes a game record. Use the cancel endpoint for soft cancellation.
      operationId: deleteGame
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Game deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the game owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/{gameId}/participation:
    post:
      tags:
        - games
      summary: Join a game
      description: Sign up for a game. If the game is full, you'll be added to the waitlist.
      operationId: joinGame
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully joined game or added to waitlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully joined the game"
                  status:
                    type: string
                    enum: [confirmed, waitlist]
                    description: Whether you were added to roster or waitlist
        '400':
          description: Invalid request or signup deadline passed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Already signed up for this game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - games
      summary: Leave a game
      description: Cancel your participation in a game. If there's a waitlist, the first person will be promoted.
      operationId: dropGame
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully left the game
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully dropped from game"
        '400':
          description: Not a participant in this game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Drop deadline has passed or game has already finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/{gameId}/cancel:
    post:
      tags:
        - games
      summary: Cancel a game
      description: |
        Cancel a game (soft delete). Sets status to 'cancelled' and records cancellation timestamp.
        Only the game owner can cancel. Cannot cancel games that have already started, finished, or completed.
        All participants will be notified (when push notifications are implemented).
      operationId: cancelGame
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Game cancelled successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the game owner or game has already started/finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notOwner:
                  value:
                    error: "Only the game owner can cancel the game"
                alreadyStarted:
                  value:
                    error: "Cannot cancel a game that has already started"
                alreadyFinished:
                  value:
                    error: "Cannot cancel a game that has already finished"
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GameCategory:
      type: string
      enum:
        - soccer
        - basketball
        - pickleball
        - flag_football
        - volleyball
        - ultimate_frisbee
        - tennis
        - other
    GameStatus:
      type: string
      enum:
        - open          # Accepting sign-ups
        - full          # Roster full, waitlist only
        - closed        # Sign-up deadline passed
        - in_progress   # Game is happening
        - completed     # Game finished
        - cancelled     # Game cancelled
...
    GameSummary:
      type: object
      description: Essential game details for list views
      properties:
        id:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/GameCategory'
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        location:
          $ref: '#/components/schemas/Location'
        startTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        maxParticipants:
          type: integer
        currentParticipants:
          type: integer
          description: Number of confirmed participants (roster, not waitlist)
        pricing:
          $ref: '#/components/schemas/Pricing'
        signupDeadline:
          type: string
          format: date-time
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, all]
        status:
          $ref: '#/components/schemas/GameStatus'
        userParticipationStatus:
          type: string
          nullable: true
          enum: [confirmed, waitlist, dropped, declined, removed]
          description: Current user's participation status (only present if user is authenticated and has joined the game)
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code for client handling
        details:
          type: object
          additionalProperties: true
          description: Additional error context